+++ crates/conversions/src/felt252.rs
@@ -186,10 +186,6 @@ impl<T: SerializeAsFelt252Vec, E: SerializeAsFelt252Vec> SerializeAsFelt252Vec f
     }
 }
 
-impl SerializeAsFelt252Vec for () {
-    fn serialize_into_felt252_vec(self, _output: &mut Vec<Felt252>) {}
-}
-
 impl<T> SerializeAsFelt252Vec for T
 where
     T: IntoConv<Felt252>,
@@ -222,3 +218,26 @@ impl SerializeAsFelt252Vec for String {
         self.as_str().serialize_as_felt252_vec()
     }
 }
+
+macro_rules! impl_serialize_for_tuple {
+    ($($ty:ident),*) => {
+        impl<$( $ty ),*> SerializeAsFelt252Vec for ( $( $ty, )* )
+        where
+        $( $ty: SerializeAsFelt252Vec, )*
+        {
+            #[allow(non_snake_case)]
+            #[allow(unused_variables)]
+            fn serialize_into_felt252_vec(self, output: &mut Vec<Felt252>) {
+                let ( $( $ty, )* ) = self;
+
+                $( $ty.serialize_into_felt252_vec(output); )*
+            }
+        }
+    };
+}
+
+impl_serialize_for_tuple!();
+impl_serialize_for_tuple!(A);
+impl_serialize_for_tuple!(A, B);
+impl_serialize_for_tuple!(A, B, C);
+impl_serialize_for_tuple!(A, B, C, D); // cairo serde supports tuples in range 0 - 4 only