+++ crates/cheatnet/src/runtime_extensions/call_to_blockifier_runtime_extension/execution/cheated_syscalls.rs
@@ -48,7 +48,7 @@ pub fn get_execution_info_syscall(
 ) -> SyscallResult<GetExecutionInfoResponse> {
     let execution_info_ptr = syscall_handler.get_or_allocate_execution_info_segment(vm)?;
 
-    let cheated_data = cheatnet_state.get_cheated_data(&syscall_handler.storage_address());
+    let cheated_data = cheatnet_state.get_cheated_data(syscall_handler.storage_address());
 
     let ptr_cheated_exec_info = get_cheated_exec_info_ptr(vm, execution_info_ptr, &cheated_data);