+++ crates/conversions/src/felt252.rs
@@ -129,11 +129,42 @@ pub trait SerializeAsFelt252Vec: Sized {
     }
 }
 
+/// use this wrapper to NOT add extra length felt
+/// useful e.g. when you need to pass an already serialized value
+pub struct RawFeltVec<T>(Vec<T>)
+where
+    T: SerializeAsFelt252Vec;
+
+impl<T> RawFeltVec<T>
+where
+    T: SerializeAsFelt252Vec,
+{
+    #[must_use]
+    pub fn new(vec: Vec<T>) -> Self {
+        Self(vec)
+    }
+}
+
+impl<T> SerializeAsFelt252Vec for RawFeltVec<T>
+where
+    T: SerializeAsFelt252Vec,
+{
+    fn serialize_into_felt252_vec(self, output: &mut Vec<Felt252>) {
+        for e in self.0 {
+            e.serialize_into_felt252_vec(output);
+        }
+    }
+}
+
 impl<T> SerializeAsFelt252Vec for Vec<T>
 where
     T: SerializeAsFelt252Vec,
 {
     fn serialize_into_felt252_vec(self, output: &mut Vec<Felt252>) {
+        let len: Felt252 = self.len().into();
+
+        len.serialize_into_felt252_vec(output);
+
         for e in self {
             e.serialize_into_felt252_vec(output);
         }
@@ -187,3 +218,26 @@ impl SerializeAsFelt252Vec for String {
         self.as_str().serialize_as_felt252_vec()
     }
 }
+
+macro_rules! impl_serialize_for_tuple {
+    ($($ty:ident),*) => {
+        impl<$( $ty ),*> SerializeAsFelt252Vec for ( $( $ty, )* )
+        where
+        $( $ty: SerializeAsFelt252Vec, )*
+        {
+            #[allow(non_snake_case)]
+            #[allow(unused_variables)]
+            fn serialize_into_felt252_vec(self, output: &mut Vec<Felt252>) {
+                let ( $( $ty, )* ) = self;
+
+                $( $ty.serialize_into_felt252_vec(output); )*
+            }
+        }
+    };
+}
+
+impl_serialize_for_tuple!();
+impl_serialize_for_tuple!(A);
+impl_serialize_for_tuple!(A, B);
+impl_serialize_for_tuple!(A, B, C);
+impl_serialize_for_tuple!(A, B, C, D); // cairo serde supports tuples in range 0 - 4 only