+++ crates/forge/src/pretty_printing.rs
@@ -1,6 +1,6 @@
-use crate::compiled_raw::CrateLocation;
 use anyhow::Error;
 use console::style;
+use forge_runner::compiled_runnable::CrateLocation;
 use forge_runner::{test_case_summary::AnyTestCaseSummary, test_crate_summary::TestCrateSummary};
 use starknet_api::block::BlockNumber;
 use std::collections::HashMap;